requirements.txt
pyannote-whisper==0.5.4
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0
ffmpeg-python
whisper @ git+https://github.com/openai/whisper.git
pyannote.audio==3.1.1
runpod==1.3.0 
